refactor: split exec resolution out into glass-exec-unix
Three crates that are not sandboxes are about to resolve external tools
(glass#374). Renaming glass-sandbox-unix wholesale would have carried
canon and dir_of — shadowed-root bind helpers — under a name that says
exec, so instead the two questions get a crate each: glass-exec-unix
answers what will run, glass-sandbox-unix answers what paths must be
exposed. Neither depends on the other.

Publishes is_executable_file, which the new importers need.

No behaviour change.
