+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["crates/glass-core", "crates/glass-x11", "crates/glass-testapp", "crates/glass-mcp", "crates/glass-wayland", "crates/glass-a11y-linux", "crates/glass-windows", "crates/glass-a11y-windows", "crates/glass-sandbox-unix", "crates/glass-sandbox-linux", "crates/glass-proc-linux", "crates/glass-clip-shim-windows", "crates/glass-dbus-linux", "crates/glass-android", "crates/glass-macos", "crates/glass-a11y-macos", "crates/glass-sandbox-macos", "crates/glass-clip-shim-macos", "crates/glass-ios"]
+members = ["crates/glass-core", "crates/glass-x11", "crates/glass-testapp", "crates/glass-mcp", "crates/glass-wayland", "crates/glass-a11y-linux", "crates/glass-windows", "crates/glass-a11y-windows", "crates/glass-exec-unix", "crates/glass-sandbox-unix", "crates/glass-sandbox-linux", "crates/glass-proc-linux", "crates/glass-clip-shim-windows", "crates/glass-dbus-linux", "crates/glass-android", "crates/glass-macos", "crates/glass-a11y-macos", "crates/glass-sandbox-macos", "crates/glass-clip-shim-macos", "crates/glass-ios"]
 # glass-fixture-egui is intentionally excluded: heavy egui deps, only exercised by #[ignore]d
 # on-box a11y tests CI can't run. Built on-demand on the box.
 exclude = ["crates/glass-fixture-egui"]
@@ -40,6 +40,7 @@ x11rb = { version = "0.14", features = ["xtest"] }
 glass-android = { path = "crates/glass-android" }
 glass-ios = { path = "crates/glass-ios" }
 glass-core = { path = "crates/glass-core" }
+glass-exec-unix = { path = "crates/glass-exec-unix" }
 glass-sandbox-unix = { path = "crates/glass-sandbox-unix" }
 glass-sandbox-linux = { path = "crates/glass-sandbox-linux" }
 glass-sandbox-macos = { path = "crates/glass-sandbox-macos" }
