+++ crates/glass-exec-unix/Cargo.toml
@@ -1,15 +1,10 @@
 [package]
-name = "glass-sandbox-macos"
+name = "glass-exec-unix"
 version = "0.0.0"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
 publish.workspace = true
-
-[dependencies]
-glass-core.workspace = true
-glass-exec-unix.workspace = true
-glass-sandbox-unix.workspace = true
 
 [dev-dependencies]
 tempfile.workspace = true
