+++ crates/glass-sandbox-linux/Cargo.toml
@@ -8,6 +8,7 @@ publish.workspace = true
 
 [target.'cfg(target_os = "linux")'.dependencies]
 glass-core.workspace = true
+glass-exec-unix.workspace = true
 glass-sandbox-unix.workspace = true
 
 [target.'cfg(target_os = "linux")'.dev-dependencies]
