+++ crates/glass-sandbox-unix/src/lib.rs
@@ -1,15 +1,15 @@
-//! Launch-target *resolution* atoms shared by glass's two sandbox backends, bwrap on Linux
-//! (`glass-sandbox-linux`) and Seatbelt on macOS (`glass-sandbox-macos`) — unix both, which is
-//! what makes this crate unix. No OS-specific containment logic lives here: only "given a
-//! program + args + cwd, what absolute host paths does the launch actually touch, resolved the
-//! way the child is exec'd." Each backend applies its OWN exposure guard/emit on top.
+//! Bind-path resolution atoms shared by glass's two sandbox backends, bwrap on Linux
+//! (`glass-sandbox-linux`) and Seatbelt on macOS (`glass-sandbox-macos`) — unix both, which is what
+//! makes this crate unix. Given a program + args + cwd, which absolute host paths does the launch
+//! actually touch. No OS-specific containment logic lives here: each backend applies its OWN
+//! exposure guard/emit on top.
 //!
-//! Everything here reads POSIX semantics — `/`-rooted paths, the execute bit, `execvp` lookup —
-//! so compiled off unix `is_absolute()` and the mode check would quietly mean something else.
+//! Whether a resolved path can be *run* is a different question, answered by `glass-exec-unix`.
+//!
+//! `is_absolute()` reads POSIX semantics, so compiled off unix it would quietly mean something else.
 #![cfg(unix)]
 #![forbid(unsafe_code)]
 
-use std::ffi::OsStr;
 use std::path::{Path, PathBuf};
 
 /// Resolve a token to an absolute host path: an absolute token as-is, a relative one against
@@ -21,31 +21,6 @@ pub fn abs_token(tok: &Path, cwd: Option<&Path>) -> Option<PathBuf> {
     } else {
         cwd.map(|c| c.join(tok))
     }
-}
-
-/// The first `$PATH` entry holding an executable regular file named `program`, resolved the way
-/// `execvp` resolves a bare command name. `None` when `$PATH` is unset or nothing matches.
-pub fn resolve_on_path(program: &OsStr) -> Option<PathBuf> {
-    let path = std::env::var_os("PATH")?;
-    resolve_on_path_in(program, &path)
-}
-
-/// [`resolve_on_path`] against an explicit `$PATH` value — the testable seam (no global env).
-/// Public so the backends' own resolution paths can be tested the same way, without a test
-/// mutating the process environment (unsound: `set_var` races any concurrent reader).
-pub fn resolve_on_path_in(program: &OsStr, path: &OsStr) -> Option<PathBuf> {
-    std::env::split_paths(path)
-        .map(|dir| dir.join(program))
-        .find(|cand| is_executable_file(cand))
-}
-
-/// Whether `p` is (or resolves through symlinks to) a regular file that is executable — `execvp`'s
-/// "is this runnable" test.
-pub(crate) fn is_executable_file(p: &Path) -> bool {
-    use std::os::unix::fs::PermissionsExt;
-    std::fs::metadata(p)
-        .map(|m| m.is_file() && m.permissions().mode() & 0o111 != 0)
-        .unwrap_or(false)
 }
 
 /// Best-effort path canonicalization that never panics on a nonexistent path: the resolved path,
@@ -67,9 +42,6 @@ pub fn dir_of(p: &Path) -> PathBuf {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use std::ffi::OsStr;
-    use std::os::unix::fs::PermissionsExt;
-    use std::path::{Path, PathBuf};
 
     #[test]
     fn abs_token_absolute_passes_through_relative_needs_cwd() {
@@ -82,47 +54,6 @@ mod tests {
             Some(PathBuf::from("/c/x/y"))
         );
         assert_eq!(abs_token(Path::new("x/y"), None), None);
-    }
-
-    #[test]
-    fn resolve_on_path_in_finds_first_executable_match() {
-        let dir = tempfile::tempdir().unwrap();
-        let exe = dir.path().join("mytool");
-        std::fs::write(&exe, b"").unwrap();
-        std::fs::set_permissions(&exe, std::fs::Permissions::from_mode(0o755)).unwrap();
-        let path = std::env::join_paths([dir.path()]).unwrap();
-        assert_eq!(resolve_on_path_in(OsStr::new("mytool"), &path), Some(exe));
-    }
-
-    /// Pins the documented "first `$PATH` entry wins" contract (`execvp` semantics): with two
-    /// directories on `$PATH`, each holding an executable of the SAME name, the match from the
-    /// FIRST directory must be returned, not merely any match.
-    #[test]
-    fn resolve_on_path_in_returns_the_first_match() {
-        let first = tempfile::tempdir().unwrap();
-        let second = tempfile::tempdir().unwrap();
-        for dir in [&first, &second] {
-            let exe = dir.path().join("mytool");
-            std::fs::write(&exe, b"").unwrap();
-            std::fs::set_permissions(&exe, std::fs::Permissions::from_mode(0o755)).unwrap();
-        }
-        let path = std::env::join_paths([first.path(), second.path()]).unwrap();
-        assert_eq!(
-            resolve_on_path_in(OsStr::new("mytool"), &path),
-            Some(first.path().join("mytool")),
-            "must return the FIRST $PATH entry's match, not merely any match"
-        );
-    }
-
-    #[test]
-    fn resolve_on_path_in_skips_non_executable_and_missing() {
-        let dir = tempfile::tempdir().unwrap();
-        let plain = dir.path().join("mytool");
-        std::fs::write(&plain, b"").unwrap();
-        std::fs::set_permissions(&plain, std::fs::Permissions::from_mode(0o644)).unwrap();
-        let path = std::env::join_paths([dir.path()]).unwrap();
-        assert_eq!(resolve_on_path_in(OsStr::new("mytool"), &path), None);
-        assert_eq!(resolve_on_path_in(OsStr::new("absent"), &path), None);
     }
 
     #[test]
@@ -144,32 +75,6 @@ mod tests {
         assert_eq!(
             canon(Path::new("/no/such/glass/path")),
             PathBuf::from("/no/such/glass/path")
-        );
-    }
-
-    #[test]
-    fn is_executable_file_true_for_exec_false_for_dir_and_plain() {
-        let dir = tempfile::tempdir().unwrap();
-
-        let exe = dir.path().join("exe");
-        std::fs::write(&exe, b"").unwrap();
-        std::fs::set_permissions(&exe, std::fs::Permissions::from_mode(0o755)).unwrap();
-        assert!(is_executable_file(&exe), "an exec-bit file must be true");
-
-        let plain = dir.path().join("plain");
-        std::fs::write(&plain, b"").unwrap();
-        std::fs::set_permissions(&plain, std::fs::Permissions::from_mode(0o644)).unwrap();
-        assert!(
-            !is_executable_file(&plain),
-            "a non-exec plain file must be false"
-        );
-
-        let subdir = dir.path().join("subdir");
-        std::fs::create_dir_all(&subdir).unwrap();
-        std::fs::set_permissions(&subdir, std::fs::Permissions::from_mode(0o755)).unwrap();
-        assert!(
-            !is_executable_file(&subdir),
-            "a directory (even an 'executable'-mode one) must be false"
         );
     }
 }
